OneDrive/Documents/South_Africa_Languanges_Translator/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.25.2
supabase==2.3.4
google-generativeai==0.8.3
pydantic==2.5.3
python-multipart==0.0.6
PyJWT==2.8.0
websockets==12.0
